Rewrite of fakegcc.py
The old implementation of fakegcc.py would call out to various
  invoke-FOO
scripts as subprocesses, giving them a path to write an XML
file to, and then parse the result.  Each script would then
invoke a side-effect tool as a subprocess (a grandchild of
the fakegcc.py process).

This was clunky, and difficult to maintain and to test.

This commit rewrites fakegcc.py to eliminate the separate
invoke-FOO scripts, instead doing them in-process as subclasses
of a new Task class.

Doing it this was allows for unit testing, which this commit
also adds.